+++ packages/mongo/polling_observe_driver.js
@@ -28,7 +28,8 @@ PollingObserveDriver = function (options) {
   // Make sure to create a separately throttled function for each
   // PollingObserveDriver object.
   self._ensurePollIsScheduled = _.throttle(
-    self._unthrottledEnsurePollIsScheduled, 50 /* ms */);
+    self._unthrottledEnsurePollIsScheduled,
+    self._cursorDescription.options.pollingThrottleMs || 50 /* ms */);
 
   // XXX figure out if we still need a queue
   self._taskQueue = new Meteor._SynchronousQueue();
@@ -43,7 +44,7 @@ PollingObserveDriver = function (options) {
         self._pendingWrites.push(fence.beginWrite());
       // Ensure a poll is scheduled... but if we already know that one is,
       // don't hit the throttled _ensurePollIsScheduled function (which might
-      // lead to us calling it unnecessarily in 50ms).
+      // lead to us calling it unnecessarily in <pollingThrottleMs> ms).
       if (self._pollsScheduledButNotStarted === 0)
         self._ensurePollIsScheduled();
     }
@@ -60,7 +61,10 @@ PollingObserveDriver = function (options) {
   if (options._testOnlyPollCallback) {
     self._testOnlyPollCallback = options._testOnlyPollCallback;
   } else {
-    var pollingInterval = self._cursorDescription.options._pollingInterval || 10 * 1000;
+    var pollingInterval =
+          self._cursorDescription.options.pollingIntervalMs ||
+          self._cursorDescription.options._pollingInterval || // COMPAT with 1.2
+          10 * 1000;
     var intervalHandle = Meteor.setInterval(
       _.bind(self._ensurePollIsScheduled, self), pollingInterval);
     self._stopCallbacks.push(function () {