+++ History.md
@@ -17,6 +17,8 @@
 
 * Allow `git+` URL schemes for npm dependencies #844
 
+* Expose options `disableOplog`, `pollingIntervalMs`, and `pollingThrottleMs` to `Cursor.find` for tuning observe parameters on the server.
+
 Patches contributed by GitHub users vereed, devgrok, ...